Move things around to clean up WookieeKafka Wookiee-Kafka no longer depends on Wookiee at all
